[40기 김채영]Add: 상수데이터 Footer 컴포넌트 구현+ Mock Data 활용하여 여러 개의 피드 구현
