+++ src/pages/chaeyoungkim/Login/Signup.js
@@ -2,15 +2,19 @@ import React, { useState } from 'react';
 // import { useNavigate } from 'react-router-dom';
 import './Login.scss';
 
-function Login() {
+function Signup() {
     // const navigate = useNavigate();
 
     // const goToMain = () => {
     //     navigate('/main-cy');
     // };
 
+    const [id, setId] = useState('');
+    const [pw, setPw] = useState('');
+    const isValid = id.includes('@') && pw.length >= 5;
+
     const handleClick = () => {
-        fetch('https://10.58.52.143:3000/auth/signin', {
+        fetch('http://10.58.52.143:3000/auth/signup', {
             method: 'POST',
             headers: { 'Content-Type': 'application/json;charset=utf-8' },
             body: JSON.stringify({ email: id, password: pw }),
@@ -20,9 +24,6 @@ function Login() {
                 console.log(result);
             });
     };
-    const [id, setId] = useState('');
-    const [pw, setPw] = useState('');
-    const isValid = id.includes('@') && pw.length >= 5;
 
     const saveUserId = event => {
         setId(event.target.value);
@@ -72,7 +73,7 @@ function Login() {
                                         : { backgroundColor: '#b0dffc' }
                                 }
                             >
-                                <span>로그인</span>
+                                <span>회원가입</span>
                             </button>
                             {/* </Link> */}
                         </div>
@@ -87,4 +88,4 @@ function Login() {
     );
 }
 
-export default Login;
+export default Signup;
