+++ src/pages/chaeyoungkim/Main/Main.scss
@@ -109,7 +109,7 @@
             border-radius: 70%;
         }
   
-        .mainArtcleUserID {
+        .mainArticleUserID {
             font-weight: bold;
             padding-left: 10px;
         }
